Refactor migration instructions into separate folder structure
Move each migration instruction to its own file for cleaner organization:
- migration-instructions/types.ts - MigrationInstruction type definition
- migration-instructions/old-profiles-location.ts - Profile location migration
- migration-instructions/registry.ts - Central registry of all migrations

The main hook now imports from the registry, making it easy to add new
migrations by creating a file and adding it to the registry.
