+++ src/installer/features/hooks/config/migration-instructions/types.ts
@@ -1,0 +1,13 @@
+/**
+ * Migration instruction type definitions
+ */
+
+/**
+ * Migration instruction interface
+ * Each migration defines a trigger function that returns:
+ * - A message string if migration is needed
+ * - null if no migration is needed
+ */
+export type MigrationInstruction = {
+  trigger: (args: { installDir: string }) => string | null;
+};
